포트찾기 시스템 강화

+++ main.js
@@ -1,6 +1,32 @@
-const { app, dialog, BrowserWindow } = require('electron');
-const { spawn, exec } = require('child_process');
+
+const { app, dialog, ipcMain, BrowserWindow } = require('electron');
+const { spawn } = require('child_process');
+
+const axios = require('axios');
+const log = require('electron-log');
 const path = require('path');
+const fs = require('fs');
+
+// clear log file
+const logFilePath = log.transports.file.getFile().path;
+
+// 렌더러에서 포트 번호 받기
+let backendPort = 8080;
+let isBackendReady = false;
+
+ipcMain.on('portnumber', (event, port) => {
+    backendPort = port;
+    isBackendReady = true;
+});
+
+if (fs.existsSync(logFilePath)) {
+    try {
+        fs.unlinkSync(logFilePath);
+        log.info("Log file cleared successfully.");
+    } catch (err) {
+        log.error("Error while clearing log file : " + err);
+    }
+}
 
 function createWindow() {
     const window = new BrowserWindow({
@@ -8,7 +34,8 @@ function createWindow() {
         height: 1000,
         webPreferences: {
             nodeIntegration: true,
-            contextIsolation: false,
+            contextIsolation: true,
+            preload: path.join(__dirname, "preload.js")
         }
     });
 
@@ -27,12 +54,11 @@ function createWindow() {
 // run new backend process
 let javaProcess = undefined;
 app.on('ready', () => {
-    const jarPath = path.join(__dirname, 'backend', 'FoxCraft.jar');
+    const jarPath = path.join(process.resourcesPath, 'FoxyCraft.jar');
     let javaExecutable = "";
-
     if (app.isPackaged) {
-        if (process.platform === "win32" && process.arch === "x64") javaExecutable = path.join(process.resourcesPath, "jar_win_64", "bin", "java.exe");
-        else if (process.platform === "darwin") javaExecutable = path.join(process.resourcesPath, "jar_mac", "Contents", "Home", "bin", "java");
+        if (process.platform === "win32" && process.arch === "x64") javaExecutable = path.join(process.resourcesPath, "jre_win_64", "bin", "java.exe");
+        else if (process.platform === "darwin") javaExecutable = path.join(process.resourcesPath, "jre_mac", "jre", "Contents", "Home", "bin", "java");
         else {
             // 지원하지 않는 OS/Architecture
             dialog.showMessageBoxSync({
@@ -45,7 +71,7 @@ app.on('ready', () => {
             return;
         }
 
-        console.log(`Starting Springboot App with ${javaExecutable}`);
+        log.info(`Starting Springboot App with ${javaExecutable}`);
         javaProcess = spawn(javaExecutable, ["-jar", jarPath]);
     } else {
         // 개발 모드에서는 gradlew run 사용
@@ -56,28 +82,34 @@ app.on('ready', () => {
         }
         // win에서는 gradlew.bat (새 창에 실행하여 로그 볼 수 있도록 하기)
         else if (process.platform === "win32") {
-            exec("cmd /c start cmd.exe /c \"cd backend && .\\gradlew.bat bootRun\"");
+            javaProcess = spawn('.\\gradlew.bat', ['bootRun'], {
+                cwd: path.join(__dirname, "backend"),
+                shell: true,
+                detached: false,
+            });
         }
     }
 
-    // console에 springboot app log출력 (mac)
+    // console에 springboot app log출력
     javaProcess?.stdout.setEncoding('utf8');
-    javaProcess?.stdout.on('data', (data) => console.log(`${data}`)).setEncoding('utf8');
-    javaProcess?.stderr.on('data', (data) => console.log(`${data}`)).setEncoding('utf8');
-    javaProcess?.on('error', (err) => console.log(`${err}`));
+    javaProcess?.stdout.on('data', (data) => log.info(`${data}`)).setEncoding('utf8');
+    javaProcess?.stderr.on('data', (data) => log.error(`${data}`)).setEncoding('utf8');
+    javaProcess?.on('error', (err) => log.error(`${err}`));
     javaProcess?.on('close', (code) => {
-        if (code !== 0) {
-            dialog.showMessageBoxSync({
-                type: "error",
-                title: "실행 중 오류 발생",
-                message: `앱 실행 중 오류가 발생했습니다. 앱 개발자에게 문의해 주십시오. Exit code : ${code}`,
-                buttons: ["확인"]
-            });
-            app.quit();
-            return;
+        if (app.isPackaged) {
+            if (code !== 0) {
+                dialog.showMessageBoxSync({
+                    type: "error",
+                    title: "실행 중 오류 발생",
+                    message: `앱 실행 중 오류가 발생했습니다. 앱 개발자에게 문의해 주십시오. Exit code : ${code}`,
+                    buttons: ["확인"]
+                });
+                app.quit();
+                return;
+            }
         }
 
-        console.log(`JavaProcess exited with code ${code}`);
+        log.info(`JavaProcess exited with code ${code}`);
     });
 
     setTimeout(() => {
@@ -85,14 +117,37 @@ app.on('ready', () => {
     }, 5000);
 });
 
-app.on('window-all-closed', () => {
-    if (process.platform !== 'darwin') app.quit();
+async function shutdown() {
+    try {
+        return await axios.post(`http://localhost:${backendPort}/actuator/shutdown`, {}, {
+            headers: {
+                "Content-Type": "application/json"
+            }
+        }).then(resp => {
+            console.log(resp.data);
+            javaProcess.kill();
+        }).catch(err => console.log(err));
+    } catch (err) {
+        console.log(err);
+    } finally {
+        if (javaProcess) javaProcess.kill();
+    }
+}
+
+let isQuitting = false; // 딱 한 번만 호출 가능하도록 설정
+app.on('before-quit', (event) => {
+    if (isQuitting) return;
+    event.preventDefault();
+    log.info("Terminating Java Process...");
+    shutdown().then(() => {
+        isQuitting = true;
+        app.quit();
+    });
 });
 
-app.on('will-quit', () => {
-    if (javaProcess) {
-        console.log("Terminating Java Process...");
-        javaProcess.kill();
+app.on('window-all-closed', () => {
+    if (process.platform !== 'darwin') {
+        app.quit();
     }
 });
